+++ CHANGES.rst
@@ -4,7 +4,11 @@ Changelog of lizard-nxt client
 Unreleased (2.4.1) (XXXX-XX-XX)
 ---------------------
 
+- Dashboard replaces time-ctx and shows omnibox cards in dashboard
+
 - Fix not compiling any translations.
+
+- Add support for viewing and deleting annotations on assets.
 
 
 Release 2.3.2 (2015-11-11)